css/css-borders-4/parsing/border-radius-bottom-computed.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Computed values of 'border-radius-bottom'</title>
<link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#corner-sizing">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<div id="target"></div>

<script>
test_computed_value("border-radius-bottom", "10px");
test_computed_value("border-radius-bottom", "10px 10px", "10px");
test_computed_value("border-radius-bottom", "5px 10px");
test_computed_value("border-radius-bottom", "10px / 5px 10px");
test_computed_value("border-radius-bottom", "5px 10px / 10px");
test_computed_value("border-radius-bottom", "10px 10px / 5px 5px", "10px / 5px");
</script>
